Changed call to malloc to use size_t as the data type so it compiles correctly under Xcode 6.2 and Xcode 6.3. Also changed logging code slightly.

+++ CalcPrimesSwift.swift
@@ -57,7 +57,7 @@ class CalcPrimesSwift:  NSObject, CalcPrimesProtocol
       var candidate: UInt32 = 3
       var isPrime: Bool = false
       let byteSize = Int(requiredComputeRecord.totalToCalculate * sizeof(UInt32))
-      var ptr = UnsafeIntPointer(malloc(UInt(byteSize)))
+      var ptr = UnsafeIntPointer(malloc(size_t(byteSize)))
       primes = UnsafeIntArray(start: ptr, count: theComputeRecord!.totalToCalculate)
       primes[primeCount++] = 2
